+++ server/impl/src/main/java/com/walmartlabs/concord/server/ServerModule.java
@@ -24,6 +24,7 @@ import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
 import com.walmartlabs.concord.db.DatabaseModule;
 import com.walmartlabs.concord.server.metrics.MetricModule;
+import com.walmartlabs.concord.server.policy.PolicyCache;
 import com.walmartlabs.concord.server.process.queue.EnqueuedTask;
 import com.walmartlabs.concord.server.task.TaskScheduler;
 
@@ -39,5 +40,6 @@ public class ServerModule extends AbstractModule {
         tasks.addBinding().to(com.walmartlabs.concord.server.process.queue.dispatcher.Dispatcher.class);
         tasks.addBinding().to(EnqueuedTask.class);
         tasks.addBinding().to(TaskScheduler.class);
+        tasks.addBinding().to(PolicyCache.class);
     }
 }