+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: cache for policies;
 - concord-server, concord-console: the "new project" and the "new
 secret" buttons can now be disabled on the organization level using
 an `entity` policy;